Expand spacewin possibilities to image and polyshape
Add new 'type's of spatial window, and support them in
spacewin_isvalid, spacewin_area and spatial_edge_correction. I did
some tests to make sure the results are close. The image type is
not quite identical to the other two, but that makes sense because
it is also not an identical window (it's cut off at pixel edges
instead of along edges of the polygons, so there are minor
discrepancies).