spacewin_gui: normalize versions that return vs save
Make sure new uicontrols are added to list of uicontrols for both versions --
the version that returns the spacewin directly to the command line as well as
the version that saves to a file or workspace variable of choice.

Also write help text, and move the copyright info down below it.

Also deal with a couple of minor linter warnings.